Put the logic in classes so I can write proper tests.
