+++ csharp-advanced-topics/ActionAndFuncDelegatesInCSharp/ActionAndFuncDelegatesInCSharp.sln
@@ -3,9 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.3.32825.248
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ActionAndFuncDelegatesInCSharp", "ActionAndFuncDelegatesInCSharp\ActionAndFuncDelegatesInCSharp.csproj", "{8C36CEA8-72E3-4939-94D2-4B418386941F}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ActionAndFuncDelegatesInCSharp", "ActionAndFuncDelegatesInCSharp\ActionAndFuncDelegatesInCSharp.csproj", "{7E91C727-9085-476E-9ABA-A2099706A99C}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Tests", "Tests\Tests\Tests.csproj", "{941FE428-5246-4F5F-84D2-0057932F9D30}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Tests", "Tests\Tests\Tests.csproj", "{A0849A3F-5BA3-4334-B87D-03C3235D17BD}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -13,19 +13,19 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{8C36CEA8-72E3-4939-94D2-4B418386941F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{8C36CEA8-72E3-4939-94D2-4B418386941F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{8C36CEA8-72E3-4939-94D2-4B418386941F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{8C36CEA8-72E3-4939-94D2-4B418386941F}.Release|Any CPU.Build.0 = Release|Any CPU
-		{941FE428-5246-4F5F-84D2-0057932F9D30}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{941FE428-5246-4F5F-84D2-0057932F9D30}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{941FE428-5246-4F5F-84D2-0057932F9D30}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{941FE428-5246-4F5F-84D2-0057932F9D30}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7E91C727-9085-476E-9ABA-A2099706A99C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7E91C727-9085-476E-9ABA-A2099706A99C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7E91C727-9085-476E-9ABA-A2099706A99C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7E91C727-9085-476E-9ABA-A2099706A99C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A0849A3F-5BA3-4334-B87D-03C3235D17BD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A0849A3F-5BA3-4334-B87D-03C3235D17BD}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A0849A3F-5BA3-4334-B87D-03C3235D17BD}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A0849A3F-5BA3-4334-B87D-03C3235D17BD}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {3D921851-2E89-457C-B5AD-BF1E7C888FBD}
+		SolutionGuid = {23C127C1-5425-491E-9B28-EEC97CAB77C2}
 	EndGlobalSection
 EndGlobal
