+++ csharp-advanced-topics/ActionAndFuncDelegatesInCSharp/ActionAndFuncDelegatesInCSharp/ActionAndFuncDelegatesInCSharp.csproj
@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>net6.0</TargetFramework>
+    <TargetFramework>net7.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
   </PropertyGroup>
